requirements.txt
arabic_reshaper==3.0.0
brisque==0.0.16
fastapi==0.111.0
numpy==1.23.5
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
python_bidi==0.4.2
ultralytics==8.2.5
dill==0.3.8